pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Site that analyzes the specified pages for SEO suitability"
authors = ["Aliev Nurlan <nualiev420@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]
repository = "https://github.com/Nurlan-Aliev/python-project-83"
classifiers = ["Topic :: Internet :: Site Management :: Link Checking"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
validators = "^0.20.0"
gunicorn = "^20.1.0"
beautifulsoup4 = "^4.12.2"
psycopg2-binary = "^2.9.6"
requests = "^2.29.0"
python-dotenv = "^1.0.0"
pook = "^1.1.1"
pytest-postgresql = "^5.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.2"
flask-testing = "^0.8.1"
pytest-cov = "^4.1.0"
pytest-postgresql = "^5.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
